+++ CHANGES.rst
@@ -9,6 +9,9 @@ Backwards-compatible changes
 * Virtualbox shared folders are disabled, in order to make Vagrant boxes
   resemble more closely nodes created by other providers.
 
+* Vagrant node instances have an extension attribute called ``ex_ssh_client``,
+  a context manager wrapping an instance of ``libcloud.compute.ssh.SSHClient``.
+
 
 Changes in version 0.3.0
 ========================